refactor: Move videosReducer in reducers folder
