+++ src/contexts/videoContext.js
@@ -1,32 +1,11 @@
 import { createContext, useState, useEffect, useContext, useReducer } from "react";
 import axios from "axios";
+import { videosReducer } from "./videoContext";
 
 const VideoContext = createContext();
 
-const videosReducer = (state, action) => {
-	switch (action.type) {
-
-	case "SET_VIDEOS_LIST":
-		return {
-			...state,
-			videos: action.payload,
-		};
-
-      case "SET_CATEGORY_LIST":
-            return {
-                  ...state,
-                  categories: action.payload,
-            };
-	default:
-		return state;
-	}
-};
-
-export {videosReducer};
-
 const VideoProvider = ({children}) => {
 
-      // const [videos, setVideos] = useState([]);
       const [loader, setLoader] = useState(false);
       const [error, setError] = useState("");
 
@@ -43,7 +22,6 @@ const VideoProvider = ({children}) => {
                   const response = await axios.get("/api/videos");
                   if(response.status === 200) {
                         setLoader(false);
-                        // setVideos(response.data.videos);
                         videosDispatch({ type: "SET_VIDEOS_LIST", payload: response.data.videos });
                   }
                   } catch(error) {
@@ -61,7 +39,6 @@ const VideoProvider = ({children}) => {
                   const response = await axios.get("/api/categories");
                   if(response.status === 200) {
                         setLoader(false);
-                        // setVideos(response.data.videos);
                         videosDispatch({ type: "SET_CATEGORY_LIST", payload: response.data.categories });
                   }
                   } catch(error) {
