+++ src/pages/VideoListing/VideoListing.jsx
@@ -4,7 +4,7 @@ import { useVideos } from "../../contexts";
 
 const VideoListing = () => {
 
-      const { videosState } = useVideos();
+      const { videosState, videosDispa} = useVideos();
 
       const { videos, categories } = videosState;
